[CHANGED]: - Code Netbeans 8.0 Java 7 Refactoring. - Put "Working" Label to "Load Media Folder" operation (Thanks Apollo FM)
[ADDED]:
- SourceDesktop Capture custom rate setting (Thanks William for the precious diff Patches) 
- StreamPanelText Local-Time-Display (Thanks Angelocracy Xue)
- StramPanelQRCode Live-Time-Display (No Streaming Software has That :D ...)
- "Lock Mixer Aspect Ratio" button in MasterPanel.
- Moving Channels Up/Down (Works only in Sequential Playlists)
- "Apply Selected Transition to All Channels" button (Thanks Apollo FM)

[FIXED]:
- ProcessRenderer HotFix. (Missed in the last build ...)
- "Apply to Stream" MasterPanel Button doesn't works with SourceText or SourceQRCode.
- Check for duplicates names adding a channel.
- "Automatic Playlist" Operation records wrong channels for the loaded stream using "Load Media  Folder".

[TODO]:
- "Playlist Editor for WS" (Thanks Apollo FM)
- fix webcam count if error writing occurs ...
- fix radioacTV effect at 720x480 ... (It's a Gstreamer issue)
- BackEnds AutoSwitch based on distro's default backends availability.
- Technically, ProcessRenderer should be modified to launch each existing
process and then Capturer should read from the InputStream of the process
instead of reading from a local TCPIP port. (Patrick Balleux)
- Adding Audio Input/Output selector (or something similar).

+++ trunk/src/com/jhlabs/image/RippleFilter.java
@@ -162,8 +162,8 @@ public class RippleFilter extends TransformFilter {
 
         @Override
 	protected void transformInverse(int x, int y, float[] out) {
-		float nx = (float)y / xWavelength;
-		float ny = (float)x / yWavelength;
+		float nx = y / xWavelength;
+		float ny = x / yWavelength;
 		float fx, fy;
 		switch (waveType) {
 		case SINE:
