+++ trunk/src/webcamstudio/sources/effects/Twirl.java
@@ -19,7 +19,7 @@ import webcamstudio.sources.effects.controls.TwirlControl;
 public class Twirl extends Effect {
 
     private final com.jhlabs.image.TwirlFilter filter = new com.jhlabs.image.TwirlFilter();
-    private float radius = (float)160f;
+    private float radius = 160f;
     private float angle = (float)Math.toRadians(90f);
     @Override
     public void applyEffect(BufferedImage img) {
